+++ webapp/src/components/HistoryUser/HistoryUser.js
@@ -1,24 +1,20 @@
-import React, { useEffect, useState } from 'react';
-import axios from 'axios';
+import React from 'react';
 import Container from '@mui/material/Container';
 import { useNavigate, useParams } from 'react-router';
 import LargeButton from '../ReactComponents/LargeButton';
 import CustomH1 from '../ReactComponents/CustomH1';
 import NavBar from "../NavBar/NavBar";
 import { useTranslation } from "react-i18next";
-import ContestRow from '../ReactComponents/ContestRow';
+import ContestList from '../ReactComponents/ContestList';
+import useFetchHistory from '../../hooks/useFetchHistory';
 
 const HistoryUser = () => {
   const { username } = useParams();
   const navigate = useNavigate();
-
-  const [contests, setContests] = useState([]);
-  const [totalTime, setTotalTime] = useState([]);
-  const [totalClues, setTotalClues] = useState([]);
-  const [numCorrect, setNumCorrect] = useState([]);
   const { t } = useTranslation();
 
   const apiEndpoint = process.env.REACT_APP_API_ENDPOINT || 'http://localhost:8000';
+  const { contests, totalTime, totalClues, numCorrect } = useFetchHistory(`${apiEndpoint}/gethistory/${username}`);
 
   const exitHistory = () => {
     navigate('/profile');
@@ -40,44 +36,6 @@ const HistoryUser = () => {
     return `${day}/${month}/${year} ${hours}:${minutes}:${seconds}`;
   };
 
-  useEffect(() => {
-    const getHistory = async () => {
-      try {
-        const response = await axios.get(`${apiEndpoint}/gethistory/${username}`);
-        setContests(response.data.contests);
-        let auxArTimes = [];
-        let auxArClues = [];
-        let auxArCorrect = [];
-        response.data.contests.forEach(contest => {
-          let auxClues = 0;
-          let auxTimes = 0;
-          let auxCorrect = 0;
-          contest.pistas.forEach(clue => {
-            auxClues += clue || 0;
-          });
-          auxArClues.push(auxClues);
-
-          contest.tiempos.forEach(time => {
-            auxTimes += time || 0;
-          });
-          auxArTimes.push(auxTimes);
-
-          contest.rightAnswers.forEach(answer => {
-            if (answer === 1) auxCorrect++;
-          });
-          auxArCorrect.push(auxCorrect);
-        });
-        setTotalClues(auxArClues);
-        setTotalTime(auxArTimes);
-        setNumCorrect(auxArCorrect);
-      } catch (error) {
-        console.error('Error al obtener el número de usuarios:', error);
-      }
-    };
-
-    getHistory();
-  }, [apiEndpoint, username]);
-
   return (
     <div>
       <NavBar />
@@ -95,29 +53,15 @@ const HistoryUser = () => {
         <LargeButton onClick={exitHistory}>
           {t("exit")}
         </LargeButton>
-        <Container
-          sx={{
-            backgroundColor: "#98d7c2",
-            marginTop: 2,
-            marginBottom: 2,
-          }}
-        >
-          <ContestRow isHeader={true} t={t} />
-          {contests.map((contest, index) => (
-            <ContestRow
-              key={contest._id || index}
-              contest={contest}
-              index={index}
-              numCorrect={numCorrect}
-              totalTime={totalTime}
-              totalClues={totalClues}
-              formatDate={formatDate}
-              enterContest={enterContest}
-              t={t}
-              isHeader={false}
-            />
-          ))}
-        </Container>
+        <ContestList
+          contests={contests}
+          numCorrect={numCorrect}
+          totalTime={totalTime}
+          totalClues={totalClues}
+          formatDate={formatDate}
+          enterContest={enterContest}
+          t={t}
+        />
       </Container>
     </div>
   );
